requirements.txt
crewai>=0.30.0
slack_sdk>=3.31.0
jira>=3.8.0
python-dotenv>=1.0.1
